refactor: separate libhijacker headers from shared contracts
source/include mixed libhijacker's private implementation headers (util.hpp
with its hand-rolled UniquePtr/String/Vector, plus kernel/, elf/, dbg/,
hijacker/) with genuinely cross-module contracts, and the whole directory was
pulled in as -isystem, so any module could reach libhijacker internals and
every warning in those headers was suppressed.

Move the libhijacker-private headers to source/libhijacker/include and expose
it as a PUBLIC include directory, so the two modules that actually use it —
util for hijacker.hpp, shellui for SHA-1 — now declare that dependency.

Split sha1.hpp: the public-domain SHA-1 goes to third_party/sha1, while
NID_KEY, Sha1 and genSha1 (which needed libhijacker's StringView) move into
the nid.hpp that uses them. shellui only ever wanted the plain digest.

Drop source/include's SYSTEM marker now that it holds only first-party
contracts. That surfaced problems -isystem had been hiding:

  * 68 format-string mismatches in dbg.hpp, printing uintptr_t with %llx
  * freebsd-helper.h redefining PROT_* and TRACEBUF over the SDK's
  * ini.h's header-only helpers declared static rather than static inline

The GNU extensions that TYPE_BEGIN/TYPE_FIELD rely on are now silenced per
header instead of by the blanket -isystem.

Also drop three redundant util.hpp includes (bootstrapper, shellui) that
referenced none of its symbols, and fix backtrace.hpp being included through
relative ../include paths.

+++ source/libhijacker/include/dbg/dbg.hpp
@@ -381,23 +381,23 @@ struct Registers : reg { // these are helpers as it is annoying that register_t 
 		rdi(a); rsi(b); rdx(c); rcx(d); r8(e); r9(f);
 	}
 	void dump() const noexcept {
-		printf("rax: 0x%08llx\n", rax());
-		printf("rbx: 0x%08llx\n", rbx());
-		printf("rcx: 0x%08llx\n", rcx());
-		printf("rdx: 0x%08llx\n", rdx());
-		printf("rsi: 0x%08llx\n", rsi());
-		printf("rdi: 0x%08llx\n", rdi());
-		printf("r8:  0x%08llx\n", r8());
-		printf("r9:  0x%08llx\n", r9());
-		printf("r10: 0x%08llx\n", r10());
-		printf("r11: 0x%08llx\n", r11());
-		printf("r12: 0x%08llx\n", r12());
-		printf("r13: 0x%08llx\n", r13());
-		printf("r14: 0x%08llx\n", r14());
-		printf("r15: 0x%08llx\n", r15());
-		printf("rbp: 0x%08llx\n", rbp());
-		printf("rsp: 0x%08llx\n", rsp());
-		printf("rip: 0x%08llx\n", rip());
+		printf("rax: 0x%08llx\n", (unsigned long long)rax());
+		printf("rbx: 0x%08llx\n", (unsigned long long)rbx());
+		printf("rcx: 0x%08llx\n", (unsigned long long)rcx());
+		printf("rdx: 0x%08llx\n", (unsigned long long)rdx());
+		printf("rsi: 0x%08llx\n", (unsigned long long)rsi());
+		printf("rdi: 0x%08llx\n", (unsigned long long)rdi());
+		printf("r8:  0x%08llx\n", (unsigned long long)r8());
+		printf("r9:  0x%08llx\n", (unsigned long long)r9());
+		printf("r10: 0x%08llx\n", (unsigned long long)r10());
+		printf("r11: 0x%08llx\n", (unsigned long long)r11());
+		printf("r12: 0x%08llx\n", (unsigned long long)r12());
+		printf("r13: 0x%08llx\n", (unsigned long long)r13());
+		printf("r14: 0x%08llx\n", (unsigned long long)r14());
+		printf("r15: 0x%08llx\n", (unsigned long long)r15());
+		printf("rbp: 0x%08llx\n", (unsigned long long)rbp());
+		printf("rsp: 0x%08llx\n", (unsigned long long)rsp());
+		printf("rip: 0x%08llx\n", (unsigned long long)rip());
 	}
 };
